+++ libraries/resource_provider.rb
@@ -227,15 +227,15 @@ module OneviewCookbook
     end
 
     # Generic method to retrieve and return a resource from different resources using the type and name of the resource
-    # @param [String, Symbol] resource_class_type Type of resource to be retrieved. e.g., :GoldenImage, :FCNetwork.
+    # @param [String, Symbol] resource_class_type Type of resource to be retrieved. e.g., :GoldenImage, :FCNetwork
     # @param [String] resource_id Name for this resource. 'Resource1'
-    # @param [String, Symbol] ret_attribute Attribute of resource to be returned by this method. Use :resource for returning complete resource.
-    def load_resource(resource_class_type, resource_id, ret_attribute = :uri)
+    # @param [NilClass, String, Symbol] ret_attribute Attribute of resource to be returned by this method. Use nil to return the complete resource
+    def load_resource(resource_class_type, resource_id, ret_attribute = nil)
       return unless resource_id
-      resource_instance = resource_named(resource_class_type).new(@item.client, name: resource_id)
-      raise "#{resource_class_type} resource with name '#{resource_id}' was not found in the appliance." unless resource_instance.retrieve!
-      return resource_instance if ret_attribute.to_sym == :resource
-      resource_instance[ret_attribute]
+      r = resource_named(resource_class_type).new(@item.client, name: resource_id)
+      raise "#{resource_class_type} '#{resource_id}' was not found in the appliance." unless r.retrieve!
+      return r unless ret_attribute
+      r[ret_attribute]
     end
   end
 end