+++ spec/unit/resources/matchers_spec.rb
@@ -240,6 +240,14 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not delete_image_streamer_deployment_plan('')
     expect(chef_run).to_not create_image_streamer_deployment_plan_if_missing('')
 
+    # image_streamer_golden_image
+    expect(chef_run).to_not create_image_streamer_golden_image('')
+    expect(chef_run).to_not delete_image_streamer_golden_image('')
+    expect(chef_run).to_not create_image_streamer_golden_image_if_missing('')
+    expect(chef_run).to_not upload_image_streamer_golden_image_if_missing('')
+    expect(chef_run).to_not download_image_streamer_golden_image('')
+    expect(chef_run).to_not download_image_streamer_golden_image_details_archive('')
+
     # image_streamer_plan_script
     expect(chef_run).to_not create_image_streamer_plan_script('')
     expect(chef_run).to_not delete_image_streamer_plan_script('')