+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.1.0
 Adds support to API300 HPE Synergy Image Streamer resources:
   - image_streamer_deployment_plan
+  - image_streamer_golden_image
 
 # 2.0.0
 Adds support to the following HPE OneView resources: